refactor(2.1): migrate clinical/ → patients/ (folder deleted)
Move all files with git mv to preserve history:

Types (clinical/types/ → patients/types/):
- clinical.ts, labs.ts, labOrders.ts
- patients/types/index.ts updated to re-export all three
- src/types/index.ts: removed stale clinical/types barrel export

Hooks (clinical/hooks/ → patients/hooks/):
- useMedications.ts, useBCMA.ts

Components (clinical/components/ → patients/components/):
- BarcodeGenerator.tsx, BarcodeScanner.tsx
- BCMAVerification.tsx, BCMAAdministration.tsx
- mar/MARModule.tsx, mar/MedicationHistoryView.tsx
- vitals/VitalsModule.tsx
- intake-output/IntakeOutputCard.tsx, AddIntakeOutputModal.tsx

Internal import fixes in moved files:
- MARModule: BBITTab import shortened to ./BBITTab (now a sibling)
- VitalsModule: 3 roundabout ../../../../features/patients/... paths → ./sibling

New barrel index files:
- patients/components/mar/index.ts (includes BBITTab + moved files)
- patients/components/vitals/index.ts
- patients/components/intake-output/index.ts

Updated 10 external importers to point to patients/:
- ModularPatientDashboard.tsx (4 imports)
- BulkLabelPrint.tsx (1 import)
- labOrderService.ts, labService.ts (2 imports)
- Labs.tsx, LabPanelDetail.tsx, LabAcknowledgeModal.tsx,
  EditLabResultModal.tsx, CreateLabPanelModal.tsx, CreateLabResultModal.tsx

Deleted: src/features/clinical/ (entire folder)

tsc --noEmit: clean ✓  lint: 14 errors (pre-existing, unchanged) ✓
