--- src/features/clinical/components/vitals/index.ts
@@ -1,1 +1,0 @@
-export { VitalsModule } from './VitalsModule';
