--- src/features/clinical/types/index.ts
@@ -1,7 +1,0 @@
-/**
- * Clinical Types - Index
- * Centralexport for all clinical-related types
- */
-
-export * from './clinical';
-export * from './labs';
