+++ src/features/patients/components/BCMAAdministration.tsx
