+++ src/features/patients/components/BarcodeGenerator.tsx
