+++ src/features/patients/components/BarcodeScanner.tsx
