+++ src/features/patients/components/CreateLabPanelModal.tsx
@@ -5,7 +5,7 @@ import React, { useState } from 'react';
 import { X, Calendar, FileText } from 'lucide-react';
 import { useTenant } from '../../../contexts/TenantContext';
 import { createLabPanel } from '../../../services/clinical/labService';
-import type { CreateLabPanelInput } from '../../../features/clinical/types/labs';
+import type { CreateLabPanelInput } from '../types/labs';
 import { getCurrentLocalDateTimeString } from '../../../utils/time';
 import { secureLogger } from '../../../lib/security/secureLogger';
