+++ src/features/patients/components/EditLabResultModal.tsx
@@ -4,7 +4,7 @@
 import React, { useState } from 'react';
 import { X } from 'lucide-react';
 import { updateLabResult } from '../../../services/clinical/labService';
-import type { LabResult } from '../../../features/clinical/types/labs';
+import type { LabResult } from '../types/labs';
 import { secureLogger } from '../../../lib/security/secureLogger';
 
 interface EditLabResultModalProps {
