+++ src/features/patients/components/intake-output/IntakeOutputCard.tsx
