+++ src/features/patients/components/intake-output/index.ts
