+++ src/features/patients/components/mar/index.ts
@@ -1,2 +1,3 @@
+export { BBITTab } from './BBITTab';
 export { MARModule } from './MARModule';
 export { MedicationHistoryView } from './MedicationHistoryView';
