+++ src/features/patients/components/vitals/index.ts
@@ -1,0 +1,5 @@
+export { VitalsModule } from './VitalsModule';
+export { VitalsContent } from './VitalsContent';
+export { VitalsTrends } from './VitalsTrends';
+export { VitalSignsEditor } from './VitalSignsEditor';
+export { NeuroAssessmentTab } from './NeuroAssessmentTab';
